refactored query methods
