+++ app/RddlInteraction/planetmint_interaction.py
@@ -2,14 +2,14 @@ import requests
 import json
 import base64
 import hashlib
-from typing import Tuple
+import binascii
 
+
+from app.dependencies import trust_wallet_instance, config
 from app.proto.planetmintgo.dao import tx_pb2 as DaoTx
 from app.proto.planetmintgo.machine import tx_pb2 as MachineTx
-from app.RddlInteraction.rddl import planetmint
-from app.RddlInteraction.rddl import signing
-from app.dependencies import trust_wallet_instance, config
-import binascii
+from app.RddlInteraction.rddl import planetmint, signing
+from app.RddlInteraction.api_queries import getAccountInfo
 
 
 def getHash(data: bytes) -> bytes:
@@ -33,60 +33,6 @@ def computeMachineIDSignature(publicKey: str) -> str:
     signature = trust_wallet_instance.sign_with_optega(2, hashBytes.hex(), publicKey)
     signature = "30" + hex(int(len(signature) / 2))[2:] + signature
     return signature
-
-
-def createAccountOnNetwork(
-    ta_service_base_url: str, machineId: str, plmnt_address: str, signature: str
-) -> requests.Response:
-    # Define the URL and data
-    url = ta_service_base_url + "/create-account"
-    data = {"machine-id": machineId, "plmnt-address": plmnt_address, "signature": signature}
-
-    # Set headers
-    headers = {"Content-Type": "application/json"}
-
-    # Send POST request with JSON data and headers
-    response = requests.post(url, json=data, headers=headers)
-    return response
-
-
-def getAccountInfo(apiURL: str, address: str) -> Tuple[int, int, str]:
-    queryURL = apiURL + "/cosmos/auth/v1beta1/account_info/" + address
-    headers = {"Content-Type": "application/json"}
-
-    # Send POST request with JSON data and headers
-    response = requests.get(queryURL, headers=headers)
-
-    accountID = 0
-    sequence = 0
-    statusMsg = ""
-    if response.status_code != 200:
-        statusMsg = response.text
-    else:
-        data = json.loads(response.text)
-        accountID = int(data["info"]["account_number"])
-        sequence = int(data["info"]["sequence"])
-        statusMsg = ""
-
-    return (accountID, sequence, statusMsg)
-
-
-def getMachineInfo(apiURL: str, address: str) -> Tuple[str, str]:
-    queryURL = apiURL + "/planetmint/machine/address/" + address
-    headers = {"Content-Type": "application/json"}
-
-    # Send POST request with JSON data and headers
-    response = requests.get(queryURL, headers=headers)
-
-    machinedata = ""
-    statusMsg = ""
-    if response.status_code != 200:
-        statusMsg = response.text
-    else:
-        machinedata = json.loads(response.text)
-        statusMsg = ""
-
-    return (machinedata, statusMsg)
 
 
 def attestMachine(
@@ -173,20 +119,6 @@ def broadcastTX(tx_bytes: str) -> requests.Response:
     print(response.status_code)
     print(response.text)
     return response
-
-
-def getBalance(address: str) -> dict:
-    url = f"{config.planetmint_api}/cosmos/bank/v1beta1/balances/{address}"
-    headers = {"Content-Type": "application/json"}
-
-    response = requests.get(url, headers=headers)
-
-    if response.status_code != 200:
-        raise Exception(f"Failed to get balance: {response.text}")
-
-    data = json.loads(response.text)
-    balance = data["balances"]
-    return balance
 
 
 def getPoPResultTx(
